+++ server/webhook-events/events.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports = Object.freeze({
+  UPDATE_CARD: 'updateCard'
+});
