+++ server/webhook-events/index.js
@@ -1,0 +1,3 @@
+module.exports = [
+  require('./update-card.js')
+];
