+++ server/webhook-events/update-card.js
@@ -1,0 +1,26 @@
+'use strict';
+const events = require('./events');
+const board = require('../board');
+const log = require('../getLogger')('update-card');
+
+module.exports = {
+  event: events.UPDATE_CARD,
+  handler: (event, sockets) => {
+    log.verbose('Card changed');
+    const card = {
+      id: event.action.data.card.id,
+      description: event.action.data.card.name
+    };
+
+    const promises = [];
+    if (event.action.data.card.idList) {
+      board.getListName(event.action.data.card.idList, process.env.TRELLO_API_TOK).then(listName => {
+        card.listID = event.action.data.card.idList;
+        card.status = listName;
+        sockets.emit('update single flight', card);
+      });
+    } else {
+      sockets.emit('update single flight', card);
+    }
+  }
+};
